Add files via upload
This dataset includes temperature data seen in India from years 1900 to 2011.It consists of a total of 111 entries and 6 columns namely- Year, Annual, Jan-Feb, Mar-May, Jun-Sept, Oct-Dec. We can see better data patterns arising in different times of the year  by looking at the various seasons in a year. This data can be used to prove Global Warming and when exactly it started showing early signs of it.
It includes line plots, heatmap (to see the correlations), box plots and distribution curve to check outliers.